requirements.txt
Pillow==11.1.0
pyusb==1.2.1
pywin32==308
libusb-package==1.0.26.3
